+++ prisma/migrations/20251210162204_add_product_disabled_flag/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Product" ADD COLUMN     "isDisabled" BOOLEAN NOT NULL DEFAULT false;